# ipx/api/handler/misc/sign_types.go
package misc

import (
	"encoding/hex"
	"errors"
	"strings"

	"github.com/andantan/evmlab/core/types"
	"github.com/andantan/evmlab/internal/util"
)

type SignRequest struct {
	Address string `json:"address" example:"0xEbD69375..."`
	Digest  string `json:"digest"  example:"0xa1b2c3d4..."`

	hash *types.Hash
}

func (r *SignRequest) ValidateRequest() error {
	r.Address = strings.TrimSpace(r.Address)
	if r.Address == "" {
		return errors.New("address is required")
	}

	b, err := util.ParseHex(r.Digest)
	if err != nil {
		return errors.New("digest: " + err.Error())
	}
	r.hash, err = types.NewHashFromBytes(b)
	if err != nil {
		return errors.New("digest: " + err.Error())
	}

	return nil
}

func (r *SignRequest) Hash() *types.Hash {
	return r.hash
}

type SignResponse struct {
	Signature string `json:"signature"`
	R         string `json:"r"`
	S         string `json:"s"`
	V         string `json:"v"`
}

func NewSignResponse(s *types.Signature) *SignResponse {
	return &SignResponse{
		Signature: "0x" + s.Hex(),
		R:         "0x" + s.R().Text(16),
		S:         "0x" + s.S().Text(16),
		V:         "0x" + s.V().Text(16),
	}
}

type EcrecoverRequest struct {
	Hash      string `json:"hash"      example:"0xa1b2c3d4..."`
	Signature string `json:"signature" example:"0xa1b2c3d4..."`

	hash *types.Hash
	sig  *types.Signature
}

func (r *EcrecoverRequest) ValidateRequest() error {
	b, err := util.ParseHex(r.Hash)
	if err != nil {
		return errors.New("hash: " + err.Error())
	}
	r.hash, err = types.NewHashFromBytes(b)
	if err != nil {
		return errors.New("hash: " + err.Error())
	}

	b, err = util.ParseHex(r.Signature)
	if err != nil {
		return errors.New("signature: " + err.Error())
	}
	r.sig, err = types.NewSignature(b)
	if err != nil {
		return errors.New("signature: " + err.Error())
	}

	return nil
}

func (r *EcrecoverRequest) ToHash() *types.Hash           { return r.hash }
func (r *EcrecoverRequest) ToSignature() *types.Signature { return r.sig }

type EcrecoverResponse struct {
	PublicKey string `json:"public_key"`
	Address   string `json:"address"`
}

func NewEcrecoverResponse(pk *types.PublicKey) *EcrecoverResponse {
	return &EcrecoverResponse{
		PublicKey: pk.Hex(),
		Address:   pk.Address().String(),
	}
}

type VerifyByPublicKeyRequest struct {
	Hash      string `json:"hash"       example:"0xa1b2c3d4..."`
	PublicKey string `json:"public_key" example:"0x04a1b2c3..."`
	Signature string `json:"signature"  example:"0xa1b2c3d4..."`

	hash   *types.Hash
	pubKey *types.PublicKey
	sig    *types.Signature
}

func (r *VerifyByPublicKeyRequest) ValidateRequest() error {
	b, err := util.ParseHex(r.Hash)
	if err != nil {
		return errors.New("hash: " + err.Error())
	}
	r.hash, err = types.NewHashFromBytes(b)
	if err != nil {
		return errors.New("hash: " + err.Error())
	}

	b, err = util.ParseHex(r.PublicKey)
	if err != nil {
		return errors.New("public_key: " + err.Error())
	}
	r.pubKey, err = types.NewPublicKeyFromBytes(b)
	if err != nil {
		return errors.New("public_key: " + err.Error())
	}

	b, err = util.ParseHex(r.Signature)
	if err != nil {
		return errors.New("signature: " + err.Error())
	}
	r.sig, err = types.NewSignature(b)
	if err != nil {
		return errors.New("signature: " + err.Error())
	}

	return nil
}

func (r *VerifyByPublicKeyRequest) ToHash() *types.Hash {
	return r.hash
}

func (r *VerifyByPublicKeyRequest) ToPublicKey() *types.PublicKey {
	return r.pubKey
}

func (r *VerifyByPublicKeyRequest) ToSignature() *types.Signature {
	return r.sig
}

type VerifyByAddressRequest struct {
	Hash      string `json:"hash"      example:"0xa1b2c3d4..."`
	Address   string `json:"address"   example:"0xAbCd1234..."`
	Signature string `json:"signature" example:"0xa1b2c3d4..."`

	hash *types.Hash
	addr *types.Address
	sig  *types.Signature
}

func (r *VerifyByAddressRequest) ValidateRequest() error {
	b, err := util.ParseHex(r.Hash)
	if err != nil {
		return errors.New("hash: " + err.Error())
	}
	r.hash, err = types.NewHashFromBytes(b)
	if err != nil {
		return errors.New("hash: " + err.Error())
	}

	b, err = util.ParseHex(r.Address)
	if err != nil {
		return errors.New("address: " + err.Error())
	}
	r.addr, err = types.NewAddressFromBytes(b)
	if err != nil {
		return errors.New("address: " + err.Error())
	}

	b, err = util.ParseHex(r.Signature)
	if err != nil {
		return errors.New("signature: " + err.Error())
	}
	r.sig, err = types.NewSignature(b)
	if err != nil {
		return errors.New("signature: " + err.Error())
	}

	return nil
}

func (r *VerifyByAddressRequest) ToHash() *types.Hash {
	return r.hash
}

func (r *VerifyByAddressRequest) ToAddress() *types.Address {
	return r.addr
}

func (r *VerifyByAddressRequest) ToSignature() *types.Signature {
	return r.sig
}

type VerifyByPublicKeyResponse struct {
	Result bool `json:"result"`
}

func NewVerifyByPublicKeyResponse(result bool) *VerifyByPublicKeyResponse {
	return &VerifyByPublicKeyResponse{Result: result}
}

type VerifyByAddressResponse struct {
	Result bool `json:"result"`
}

func NewVerifyByAddressResponse(result bool) *VerifyByAddressResponse {
	return &VerifyByAddressResponse{Result: result}
}

type SignLegacyTransactionRequest struct {
	Address     string `json:"address"      example:"0xEbD69375..."`
	UnsignedRLP string `json:"unsigned_rlp" example:"0xec8085..."`

	unsignedRaw []byte
}

func (r *SignLegacyTransactionRequest) ValidateRequest() error {
	r.Address = strings.TrimSpace(r.Address)
	if r.Address == "" {
		return errors.New("address is required")
	}

	b, err := util.ParseHex(r.UnsignedRLP)
	if err != nil {
		return errors.New("unsigned_rlp: " + err.Error())
	}
	if len(b) == 0 {
		return errors.New("unsigned_rlp: must not be empty")
	}
	r.unsignedRaw = b

	return nil
}

func (r *SignLegacyTransactionRequest) UnsignedRaw() []byte { return r.unsignedRaw }

type SignLegacyTransactionResponse struct {
	RawTransaction string `json:"raw_transaction"`
	TxHash         string `json:"tx_hash"`
	R              string `json:"r"`
	S              string `json:"s"`
	V              string `json:"v"`
}

func NewSignLegacyTransactionResponse(raw []byte, hash *types.Hash, sig *types.Signature) *SignLegacyTransactionResponse {
	return &SignLegacyTransactionResponse{
		RawTransaction: "0x" + hex.EncodeToString(raw),
		TxHash:         hash.String(),
		R:              "0x" + sig.R().Text(16),
		S:              "0x" + sig.S().Text(16),
		V:              "0x" + sig.V().Text(16),
	}
}

type SignEIP1559TransactionRequest struct {
	Address     string `json:"address"      example:"0xEbD69375..."`
	UnsignedRLP string `json:"unsigned_rlp" example:"0x02f8..."`

	unsignedRaw []byte
}

func (r *SignEIP1559TransactionRequest) ValidateRequest() error {
	r.Address = strings.TrimSpace(r.Address)
	if r.Address == "" {
		return errors.New("address is required")
	}

	b, err := util.ParseHex(r.UnsignedRLP)
	if err != nil {
		return errors.New("unsigned_rlp: " + err.Error())
	}
	if len(b) == 0 {
		return errors.New("unsigned_rlp: must not be empty")
	}
	r.unsignedRaw = b

	return nil
}

func (r *SignEIP1559TransactionRequest) UnsignedRaw() []byte {
	return r.unsignedRaw
}

type SignEIP1559TransactionResponse struct {
	RawTransaction string `json:"raw_transaction"`
	TxHash         string `json:"tx_hash"`
	R              string `json:"r"`
	S              string `json:"s"`
	V              string `json:"v"`
}

func NewSignEIP1559TransactionResponse(raw []byte, hash *types.Hash, sig *types.Signature) *SignEIP1559TransactionResponse {
	return &SignEIP1559TransactionResponse{
		RawTransaction: "0x" + hex.EncodeToString(raw),
		TxHash:         hash.String(),
		R:              "0x" + sig.R().Text(16),
		S:              "0x" + sig.S().Text(16),
		V:              "0x" + sig.V().Text(16),
	}
}