# ipx/docs/swagger.yaml
basePath: /
definitions:
...
  misc.EcrecoverRequest:
    properties:
      hash:
        example: 0xa1b2c3d4...
        type: string
      signature:
        example: 0xa1b2c3d4...
        type: string
    type: object
  misc.EcrecoverResponse:
    properties:
      address:
        type: string
      public_key:
        type: string
    type: object
...
  misc.SignEIP1559TransactionRequest:
    properties:
      address:
        example: 0xEbD69375...
        type: string
      unsigned_rlp:
        example: 0x02f8...
        type: string
    type: object
  misc.SignEIP1559TransactionResponse:
    properties:
      r:
        type: string
      raw_transaction:
        type: string
      s:
        type: string
      tx_hash:
        type: string
      v:
        type: string
    type: object
  misc.SignLegacyTransactionRequest:
    properties:
      address:
        example: 0xEbD69375...
        type: string
      unsigned_rlp:
        example: 0xec8085...
        type: string
    type: object
  misc.SignLegacyTransactionResponse:
    properties:
      r:
        type: string
      raw_transaction:
        type: string
      s:
        type: string
      tx_hash:
        type: string
      v:
        type: string
    type: object
  misc.SignRequest:
    properties:
      address:
        example: 0xEbD69375...
        type: string
      digest:
        example: 0xa1b2c3d4...
        type: string
    type: object
  misc.SignResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signature:
        type: string
      v:
        type: string
    type: object
...
  misc.VerifyByAddressRequest:
    properties:
      address:
        example: 0xAbCd1234...
        type: string
      hash:
        example: 0xa1b2c3d4...
        type: string
      signature:
        example: 0xa1b2c3d4...
        type: string
    type: object
  misc.VerifyByAddressResponse:
    properties:
      result:
        type: boolean
    type: object
  misc.VerifyByPublicKeyRequest:
    properties:
      hash:
        example: 0xa1b2c3d4...
        type: string
      public_key:
        example: 0x04a1b2c3...
        type: string
      signature:
        example: 0xa1b2c3d4...
        type: string
    type: object
  misc.VerifyByPublicKeyResponse:
    properties:
      result:
        type: boolean
    type: object
  v1.BuildEIP1559TransactionRequest:
    properties:
      chain_id:
        example: "20001209"
        type: string
      data:
        example: 0x
        type: string
      gas_limit:
        example: 21000
        type: integer
      max_fee_per_gas:
        example: "3000000000"
        type: string
      max_priority_fee_per_gas:
        example: "1500000000"
        type: string
      nonce:
        example: 0
        type: integer
      to:
        example: 0x8336c196ABb9E7092C879C28D352b39d3f2f3D7A
        type: string
      value:
        example: "1000000000000000000"
        type: string
    type: object
  v1.BuildEIP1559TransactionResponse:
    properties:
      signing_hash:
        type: string
      unsigned_rlp:
        type: string
    type: object
  v1.BuildLegacyTransactionRequest:
    properties:
      chain_id:
        example: "20001209"
        type: string
      data:
        example: 0x
        type: string
      gas_limit:
        example: 21000
        type: integer
      gas_price:
        example: "20000000000"
        type: string
      nonce:
        example: 0
        type: integer
      to:
        example: 0x8336c196ABb9E7092C879C28D352b39d3f2f3D7A
        type: string
      value:
        example: "1000000000000000000"
        type: string
    type: object
  v1.BuildLegacyTransactionResponse:
    properties:
      signing_hash:
        type: string
      unsigned_rlp:
        type: string
    type: object
...
host: localhost:33152
info:
  contact: {}
  description: Ethereum transaction API
  title: evmlab API
  version: "1.0"
paths:
...
  /evm/sign:
    post:
      consumes:
      - application/json
      description: Signs a hex-encoded 32-byte digest with the given private key using
        raw secp256k1 ECDSA. Hashing is the caller's responsibility — use the hash
        endpoints to produce the digest before calling this.
      parameters:
      - description: Address and digest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.SignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.SignResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign a pre-computed hash
      tags:
      - sign
  /evm/sign/ecrecover:
    post:
      consumes:
      - application/json
      description: Recovers the uncompressed secp256k1 public key and Ethereum address
        from a 32-byte hash and a recoverable ECDSA signature in [R || S || V] format
      parameters:
      - description: Hash and signature
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.EcrecoverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.EcrecoverResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Recover public key and address from a signature
      tags:
      - sign
  /evm/sign/transaction/eip1559:
    post:
      consumes:
      - application/json
      description: Decodes an unsigned EIP-1559 payload, signs it with the key for
        the given address, and returns the signed raw transaction and tx hash
      parameters:
      - description: Address and unsigned RLP
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.SignEIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.SignEIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign an unsigned dynamic fee native transfer transaction
      tags:
      - transaction
  /evm/sign/transaction/legacy:
    post:
      consumes:
      - application/json
      description: Decodes an unsigned EIP-155 legacy RLP, signs it with the key for
        the given address, and returns the signed raw transaction and tx hash
      parameters:
      - description: Address and unsigned RLP
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.SignLegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.SignLegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign an unsigned legacy transaction
      tags:
      - sign
  /evm/sign/verify/by-address:
    post:
      consumes:
      - application/json
      description: Recovers the signer's address from the signature via ecrecover
        and compares it against the provided address
      parameters:
      - description: Hash, address, and signature
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.VerifyByAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.VerifyByAddressResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify a signature against an address
      tags:
      - sign
  /evm/sign/verify/by-public-key:
    post:
      consumes:
      - application/json
      description: Recovers the signer's public key from the signature via ecrecover
        and compares it against the provided public key
      parameters:
      - description: Hash, public key, and signature
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.VerifyByPublicKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.VerifyByPublicKeyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify a signature against a public key
      tags:
      - sign
...
  /evm/v1/transaction/eip1559/build:
    post:
      consumes:
      - application/json
      description: Constructs an unsigned EIP-1559 transaction for native ETH transfer
        and returns the encoded signing payload and signing hash
      parameters:
      - description: Transaction fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.BuildEIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.BuildEIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build an unsigned dynamic fee native transfer transaction
      tags:
      - transaction
  /evm/v1/transaction/legacy/build:
    post:
      consumes:
      - application/json
      description: Constructs an unsigned EIP-155 legacy transaction for native ETH
        transfer and returns the RLP encoding and signing hash
      parameters:
      - description: Transaction fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.BuildLegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.BuildLegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build an unsigned legacy native transfer transaction
      tags:
      - transaction
...
swagger: "2.0"